feat: Use gitlab.localtest.me domain to unify access
The gitlab.localtest.me will resolve to the gitlab container inside
the docker network created by the docker-compose. It will resolve
to localhost from the host. This will allow using the same url
to access GitLab UI, clone the repo and reach the API endpoints
from the GitLab CI jobs.